+++ src/routes/examples/RenderItem.svelte
@@ -1,7 +1,0 @@
-<script>
-	export let item
-</script>
-
-<div>
-	{item.$selected ? '👌' : '👉'} {item.text} {item.$selected ? '✅' : '☑️'}
-</div>